Move state objects back to `redux/reducers`
