+++ client/src/redux/actions.ts
@@ -12,34 +12,6 @@ import {HvError} from '../typeDefinitions'
 import {handleAction} from './utils'
 import {push} from 'react-router-redux'
 
-// This interface defines types of the entire state object of Hollowverse
-export interface IGeneralState {
-  searchInputValue: string,
-  searchResults: IAlgoliaSearchResults | undefined,
-  isSearchPending: boolean,
-  loginStatus: facebookSdk.LoginStatus,
-  isLoginPending: boolean,
-  isLogoutPending: boolean,
-  error: HvError,
-  isNavMenuOpen: boolean,
-  lastSearchTerm: string,
-  createProfileUrlInputValue: string
-}
-
-// Initialize the default state object
-export const initialGeneralState: IGeneralState = {
-  searchInputValue: '',
-  searchResults: undefined,
-  isSearchPending: false,
-  loginStatus: 'unknown',
-  isLoginPending: false,
-  isLogoutPending: false,
-  error: undefined,
-  isNavMenuOpen: false,
-  lastSearchTerm: '',
-  createProfileUrlInputValue: '',
-}
-
 // These are all the actions that can be triggered from within the Hollowverse application
 export const actions = {
   navigateToSearch: (payload: string) => {
@@ -58,11 +30,13 @@ export const actions = {
   setError: (payload: HvError) => ({type: 'setError', payload}),
   setIsNavMenuOpen: (payload: boolean) => ({type: 'setIsNavMenuOpen', payload}),
   setLastSearchTerm: (payload: string) => ({type: 'setLastSearchTerm', payload}),
+  setCreateProfileUrlInputValue: (payload: string) => ({type: 'setCreateProfileUrlInputValue', payload}),
+
+  // Redux Sagas actions
   requestSearchResults: (payload: string) => ({type: 'requestSearchResults', payload}),
   requestLogin: () => ({type: 'requestLogin'}),
   requestLogout: () => ({type: 'requestLogout'}),
   requestUpdateLoginStatus: () => ({type: 'requestUpdateLoginStatus'}),
-  setCreateProfileUrlInputValue: (payload: string) => ({type: 'setCreateProfileUrlInputValue', payload}),
 }
 
 export type ActionTypes = keyof typeof actions
